+QuestionParsers which selects the parser; *nicer propose answer message; +default experience settings in qukkiz.yml; !use exp reward; !list hinter hints a random visible answer; !ConfigurationParser reads text questions and aliased answers correct; *answer delta is now a double; +AnswerResult.INVALIDATED to define answers which were valid but already mentioned; !ListQuestion concat the visible answers of all answers; +experience rewarded message;

+++ src/main/java/de/xzise/qukkiz/questions/QuestionInterface.java
@@ -11,14 +11,13 @@ public interface QuestionInterface {
      *            Tested answer.
      * @return Following values could be returned:
      *         <ul>
-     *         <li>null – the answer is incorrect and impossible</li>
-     *         <li>{@link Integer#MAX_VALUE}/{@link Integer#MIN_VALUE} – the
-     *         answer was incorrect but possible</li>
+     *         <li>{@link Double#NaN} – the answer is incorrect and impossible</li>
+     *         <li>{@link Double#NEGATIVE_INFINITY} / {@link Double#POSITIVE_INFINITY} – the answer was incorrect but possible</li>
      *         <li>0 – the answer is perfect.</li>
      *         <li>Otherwise it define the difference to the answer.</li>
      *         </ul>
      */
-    Integer testAnswer(String answer);
+    double testAnswer(String answer);
 
     String getAnswer();
